refactored SingleBlog and FullBlog into 1 compnent
